rename secir vaccine files
