+++ cpp/models/secir_vaccine/parameters.h
